+++ src/remembrall/helpers/customErrors.py
@@ -1,6 +1,3 @@
-# NOTE: We shjould get rid of this file and
-# move exceptions to helpers folder to avoid circular imports
-
 class ShortName(Exception):
     pass
