+++ tests/test_book_operations.py
@@ -1,15 +1,16 @@
 
 import unittest
 from datetime import datetime, timedelta
-from remembrall.helpers.book_operations import *
+
+import remembrall.helpers.book_operations as sut
 
 class TestBookOperations(unittest.TestCase):
 
     def setUp(self):
-        self.book = AddressBook()
-        self.record1 = Record("Dmytro")
+        self.book = sut.AddressBook()
+        self.record1 = sut.Record("Dmytro")
         self.record1.add_phone("1234567890")
-        self.record2 = Record("Ivan")
+        self.record2 = sut.Record("Ivan")
         self.record2.add_phone("9876543210")
         self.record2.add_birthday("28.02.1988")
         self.book.add_record(self.record1)
@@ -17,22 +18,22 @@ class TestBookOperations(unittest.TestCase):
 
     def test_add_contact(self):
         args = ["Artem", "5555555555"]
-        add_contact(args, self.book)
+        sut.add_contact(args, self.book)
         self.assertIn("Artem", self.book.data)
 
     def test_change_contact(self):
         args = ["Dmytro", "Anton"]
-        change_contact(args, self.book)
+        sut.change_contact(args, self.book)
         self.assertEqual(str(self.record1.name), "Anton")
 
     def test_remove_contact(self):
         args = ["Ivan"]
-        remove_contact(args, self.book)
+        sut.remove_contact(args, self.book)
         self.assertNotIn(self.record2, self.book.data.values())
 
     def test_change_phone(self):
         args = ["Ivan", "9876543210", "9999999999"]
-        change_phone(args, self.book)
+        sut.change_phone(args, self.book)
 
         phones = [phone.value for phone in self.record2.phones]
         self.assertIn("9999999999", phones)
@@ -40,24 +41,24 @@ class TestBookOperations(unittest.TestCase):
 
     def test_remove_phone(self):
         args = ["Dmytro", "1234567890"]
-        remove_phone(args, self.book)
+        sut.remove_phone(args, self.book)
         self.assertTrue(len(self.record1.phones) == 0)
 
     def test_get_contact_phones_single(self):
         args = ["Dmytro"]
-        result = get_contact_phones(args, self.book)
+        result = sut.get_contact_phones(args, self.book)
         self.assertEqual(result, "1234567890")
 
     def test_get_contact_phones_multiple(self):
         args = ["Dmytro", "5555555555"]
-        add_contact(args, self.book)
+        sut.add_contact(args, self.book)
 
         args = ["Dmytro"]
-        result = get_contact_phones(args, self.book)
+        result = sut.get_contact_phones(args, self.book)
         self.assertEqual(result, "1234567890; 5555555555")
 
     def test_get_all_contacts(self):
-        result = get_all_contacts(self.book)
+        result = sut.get_all_contacts(self.book)
         self.assertEqual(result,
                         "Address book:\n" \
                         "Contact name: Dmytro, phones: 1234567890, birthday: \n" \
@@ -66,17 +67,17 @@ class TestBookOperations(unittest.TestCase):
 
     def test_add_birthday(self):
         args = ["Dmytro", "01.04.1990"]
-        add_birthday(args, self.book)
+        sut.add_birthday(args, self.book)
         self.assertEqual(str(self.record1.birthday), "01.04.1990")
 
     def test_show_birthday(self):
         args = ["Ivan"]
-        result = show_birthday(args, self.book)
+        result = sut.show_birthday(args, self.book)
         self.assertEqual(str(result), "28.02.1988")
 
     def test_change_birthday(self):
         args = ["Ivan", "25.02.1988"]
-        change_birthday(args, self.book)
+        sut.change_birthday(args, self.book)
         self.assertEqual(str(self.record2.birthday), "25.02.1988")
 
     def test_get_birthdays(self):
@@ -91,9 +92,9 @@ class TestBookOperations(unittest.TestCase):
             birthday = today + timedelta(days=1)
 
         args = ["Dmytro", f"{birthday.day}.{birthday.month}.1995"]
-        add_birthday(args, self.book)
+        sut.add_birthday(args, self.book)
 
-        result = get_birthdays([7], self.book)
+        result = sut.get_birthdays([7], self.book)
         self.assertEqual(result,
                          "Upcoming birthdays:\n" \
                          f"{{'name': 'Dmytro', 'congratulation_date': '{birthday.strftime("%d.%m.%Y")}'}}"
